DPDK: Fix source for tarball (#3505)
* urlparse tar fix

Fixing one bug uncovered a few others.
Fixes the way the Tar tool handles fetching
the filename of the tar file it downloads.

* Wget: add caching

* Tar: Allow skip-old-files option

When running Wget.get(..., force_run=False) and Tar.extract it is
useful to allow Tar to skip extracting existing files on the second
pass.

Allow the skip-old-files option, so Tar.extract will not overwrite
existing files in the output directory.

Note: it's important to not use this option if you are providing LISA
with a default filename for your tarballs. This option could silently
allow Tar.extract to not update the contents of a directory with the
newer file contents.

I don't see anyone using that schema now. My apologies to future devs who
find this commit message while debugging that issue.

Tar: chi comments

* TarDownloader: enable skip_old_files option in tar extract

Dpdk: chi comments

* Tar: fix subclass prototype

* Wget: remove unneeded return