Process: fix found position bug with wait_output delta_only
Fixes: 603bd277196f8

If a buffer is length N and found message 0 is not at
buffer[N-len(message0):N], the current code can result in missed
messages on future calls to wait_output.

wait_output currently stores the total length of the buffer as
the search buffer offset, even when a keyword is found.

Consider a buffer of 100 characters, keyword of len 10 found at 50.
The old code will record 100 as the last length. So if your next
keyword is length 10 and found at position 70, it will be lost
if you call wait_output with delta_only again.

This commit fixes that issue by saving only the end position
of slice buffer[start:end] which contains the first keyword.

Otherwise, it saves the length of the buffer if no keyword was
found.

Signed-off-by: Matthew G McGovern <mamcgove@microsoft.com>