Process: fix found position bug with wait_output delta_only
wait_output stores the total length of the string as the end position,
even when a message is found.

If a buffer is length N and found message 0 is not at
buffer[N-len(message0):N], the current code can result in missed
messages on future calls to wait_output.

Consider a buffer of 100 characters, keyword of len 10 found at 50.
The old code will record 100 as the last length. So if your next message
is length 10 and found at position 70, it will be lost if you call
delta_only again.

This commit fixes that issue by saving only the end position
of slice buffer[start:end] which contains the first keyword.

Signed-off-by: Matthew G McGovern <mamcgove@microsoft.com>