ADOArtifactsDownloader: make output path configurable
Keep the defaults prior to the change, but make it so that an user is
able to fine tune the destination (e.g. use a cache dir, to avoid
costly re-downloads of the same thing).

In order to make caching possible, also put build_id in the namespace
of the downloaded asset.