ado_artifact_download: add extra artifact field/property filtering
The method/helper will now get a new filtering mechanism, beyond the

  if run.result == "succeeded" and run.state == "completed"

hardcoded one.

Imagine one needs to check for the presence of the following
attributes, with the following values:

    additional_build_checks = {
        "additional_properties" : {
            "templateParameters": {
                "build_tools_image": True,
                "build_base_os": True
    	    }
        }
    }

With this commit, it is now possible. Without it, it would blindly
pick the last succesful build in the pipeline, that could still miss
artifacts we are interested in.