Introduce check for BTF sysfs
The BTF sysfs interface is important because it provides a standardized way for
the kernel to expose BPF Type Format (BTF) data to user space, enabling
advanced introspection of kernel and module types. This accessibility is
crucial for BPF tooling, such as bpftool and libbpf, to perform type-aware
operations, debugging, and verification.The BTF sysfs files are at
/sys/kernel/btf/ by standard, where each entry corresponds to BTF data for the
kernel or loaded kernel modules.

Add bpf testsuite which can be expanded with more tests