tools: make expected exit code optional for cargo test
Subtest failure of `verify_rust_vmm_mshv_tests` results in an
assertion error for the entire test case. The cargo tool
implementation for executing the test itself asserts a non zero
exit code & results in cargo test failure which hides the subtest
failure.

To control whether the test should assert on command exit code,
make the exit code params optional. This ensures that the subtests
are individually reported as failed instead of throwing an assert
error on the entire test case.

Signed-off-by: AASTHA RAWAT <aastharawat@microsoft.com>