Add new testsuite for firewalld (#3882)
* Adds new testsuite for firewalld

This commit adds a new testsuite for firewalld(a firewall daemon with d-bus interface) in LISA.
The suite runs the compiled testsuite binary provided by the firewalld-test rpm.

- the test is supported only for AzureLinux 3.0.
- there are known test failures because of missing kernel configs.
  the test checks the failure list against this known list for setting status.

* fix flake8 & mypy errors

* Add proper logs

* Address review comments

* Update failure logic to account for aarch64 runs.

Aarch64 kernel has additional FIB configs enabled compared to
x86_64 kernel. The failure list is thus a subset of the known
list in x86_64.

* Fix test description

* Remove setting test result status and log the test numbers