XFSTest tool hostfix - for Debian 12 and other distros (#3906)
* Update xfstesting.py

Bug fix for xfstesting.py, missing tab in test list is causing xfstest for SMB to fail.

* add fix for legacy OS distros

add fixes for legacy Distros that are not able to build newer release of XFStest. In this case we fall back to last known Tag that can build on the respective distros without requirement of additional or "out of band" packages from non-official repos.

* additional cleanup, compliance with flake8 and black formatting

* Update xfstests.py

* Update xfstests.py

* updated lookup to use node.os.information

* formatting updates

* added comments , minor syntax fixes and changed default for SLES 12- SP5

* updated xfstests to vendor_release format

* Update xfstests.py

* Update xfstests.py

* Update xfstests.py

* updating XFStests.py with hotfix to cover more distributions that are limited by XFStests release tag

* Update xfstests.py

* Updated XFStests.py to account for RHEL,/ CentOs < 7.0

Fixed a bug where incorrect GCC softlink to an non-existing executable was being created during package installation, causing make to fail for XFSTests tool.

* removed unused import