Allow runbook to override image requirements
Make all image requirements default to `None`. This allows LISA to differentiate between when a value is provided in the runbook or not. If the requirement is None, then there will be no change to the capability when it is merged.

If image requirements are provided in the runbook, do not override them with the automatically queried requirements.