dpdk: force recent stable default source on mana
MANA is new and using the package manager turns out to
be less reliable than just building an LKG version of DPDK
and rdma-core.

This change forces a default source build for that hardware.
This will help stabilize issues where tests need to install
then re-install dpdk from both apt and source on mana, as well.