DPDK: install path fixes for meson and Ubuntu 24.04 (#3598)
* Meson: try both 'meson' and 'python3-meson'

* DPDK: Add needs_reboot option to installer

* Suse: implement _uninstall_packages

* Suse: i guess just 0

* Meson: restructure and comment installation options

* package_exists: add minimum_version

* rdma-core: check for existence before removal

* Dpdk: move pyelftools to single installation point

* Minimum version add debug log if no match

* OS SUSE uninstall: use assert_that

* OS packages, fix printing of List[str] of packages

* make package_exists version check simpler

* DPDK: prefer newer pyelftools package

* Use helper function instead of instantiating the class directly

* check_exists( ..., minimum_version) considered harmful

Undoing a change I added which enabled a _minimum_version check on
_check_exists. This is a bad system because you'll end up needing to do
things like this:

if not posix_os._check_exists(pkg, minimum_version=...):
	# Now what? Was it installed or not?
	# Was it installed and just the wrong version?

The ultimate fix should be is_package_in_repo(pkg, minimum_version)
Since we should be checking if an available package is the right version
_before_ installing it blindly.

This all seems out-of-scope for fixing a bug installing meson
in _just_ the DPDK suite. So, I'm removing it. We can move the fixed
install pattern to the operating_system.py level later if anyone else
needs it.

* DPDK: linux-modules-extra-azure reboot

* Meson: rework install flow, plus more and better comments.

* fix VersionInfo -> LisaVersionInfo mismatch

* meson: compact if installed logic

* meson: edit some comments for clarity