Move examples and microsoft directories into the Python package (#4023)
* Move examples and microsoft into Python package

This moves the examples and microsoft directories in the repository root
into the lisa Python package. Moving these inside the package allows
them to be distributed via sdist and wheels and since the lisa package
references these files, prior to this commit installs from sdist/wheels
did not work.

* Rename package from lisa to mslisa

The project is still importable as lisa, but the Python package is named
"mslisa".

* Extensions: Add compatibility for legacy paths.

Due to packaging changes, existing runbook paths
may be incorrect. This update ensures legacy paths
remain functional. Tests are still imported from
`examples` or `microsoft` within the new packaging
structure for consistency.

* Runbook: Conditional import of builtin tests

Due to packaging changes, the test suite folder
may be harder to locate outside the `lisa` folder.
To simplify importing, a flag is added:

- Introduce `import_builtin_tests` boolean field
  in the Runbook schema, defaulting to `False`.

* Drop symlinks and tweak the old path fixup

- The symlinks cause more trouble than they're worth when combined with
  this fixup function since apparently symlinks on Windows are weird.

- The path fixup code should build the path from the repo root to
  account for nested directories.

* nox: fix config for isort and mypy

microsoft and examples folders are now in lisa
proper.

1. Drop microsoft and examples dirs from pylint.
1. Ignore dev folders for isort and mypy.
2. Remove microsoft folder from mypy.

* fix isort issue after moved folders

* fix mypy issues after moved folders

---------

Co-authored-by: Chi Song (from Dev Box) <chisong@microsoft.com>