dpdk: multiple senders single receiver muliple port testpmd
Adds a test with multiple senders and a single receiver
using two ports.

The largest change for this requires fixing the argument generation
and handling of which nics to use. Now, nics are picked
explicitly by subnet to ensure senders and receivers
are able to reach each other. Nic name and list order
are not reliable.

The other change is passing a list of NicInfo objects to the
argument generation functions instead of relying on an implicit
default. Now, multiple ports can be included arbitrarily.

Signed-off-by: Matthew McGovern (LINUX) <mamcgove@microsoft.com>