tools: add arg to control failure in get_tag
get_tag() throws an error whenever the current commit doesn't correspond
to a tag. This is undesirable. There could be cases where we want to use
untagged commits. For example, running kernel installer on an arbitrary
kernel branch (instead of official one).

Add an argument to get_tag() that controls if the function fails when
not tag is found. Set it to true by default to accommodate existin
users.

For kernel installer, pass it as false.

Signed-off-by: Anirudh Rayabharam <anrayabh@microsoft.com>