SecurityProfile for Windows Images
Windows and Linux images both use the tag "TrustedLaunchAndConfidentialVmSupported" but their use of the tag is different.

Windows images are generally GP with CVM support. They are intended to be used on GP hardware or CVM hardware. It is a single combined image for both scenarios.

Linux images are generally intended only for CVM. There are separate images for GP scenarios. Testing Linux CVM images on GP hardware is generally not testing an intended use case.

This change treats Windows and Linux VMs separately to recognize this difference in labeling conventions.