Include cmd id in error message
Processes may be running in parallel. This can make it difficult to correlate the exit code error message with the process output logs above.

Including the process id in the error message makes it easier to know which command output is relevant to the process with a unexpected exit code.