remote_copy: Add method to scp file[s] from src to dest node
`copy_between_remotes` method does scp of a file[s] (recursively) from source
node to destination node. The scp command is executed on the source node. If the
destination node is an azure vm, make sure to add the IP address of the source
node in the list of `source_address_prefixes` of the destination node for the
command to succeed.

Signed-off-by: Piyush Sachdeva <psachdeva@microsoft.com>
Signed-off-by: Piyush Sachdeva <s.piyush1024@gmail.com>